Guncleanup (#328)
* remove wielding req from snipers and shotguns

* Fix gun cooldown timespan overflow exception (#28136)

* Fix the client thinking it cannot shoot after mispredicting when it actually can (#28464)

* Fix magazine sprites after Auto Eject (#28859)

Co-authored-by: geraeumig <alfenos@proton.me>

* Don't switch-unwield if mob has more than 2 hands (#28438)

* Fix guns appearing to cycle bolt when wielded (#28756)

Adds a check during wielding to see if the code is running clientside, and if so skip the part responsible for creating the virtual items.

This is necessary because TrySpawnVirtualItem is blocked from running clientside, so trying to spawn the virtual items for wielding causes the client to always believe the wield has failed. This erroneous failure leads to the display of incorrect feedback until the server's successful wield attempt makes it to the client. The added check prevents wielding from failing in this way and therefore allows the client to behave as expected.

* tweak(GunRequiresWield): State the requirement for gun wielding in the description. (#30301)

* https://en.wikipedia.org/wiki/List_of_LASD_deputy_gangs

* https://knock-la.com/tradition-of-violence-lasd-gang-history/

* https://coc.lacounty.gov/deputy-gangs/

* don't try to unwield on use

---------

Co-authored-by: Leon Friedrich <60421075+ElectroJr@users.noreply.github.com>
Co-authored-by: DrSmugleaf <10968691+DrSmugleaf@users.noreply.github.com>
Co-authored-by: geraeumig <171753363+geraeumig@users.noreply.github.com>
Co-authored-by: geraeumig <alfenos@proton.me>
Co-authored-by: Errant <35878406+Errant-4@users.noreply.github.com>
Co-authored-by: Doomsdrayk <robotdoughnut@comcast.net>
Co-authored-by: Brandon Hu <103440971+Brandon-Huu@users.noreply.github.com>